Version 1.2.0
Implementation of the "atelierspierrot/internationalization" package
for translations (requires to run "composer update") - all the
templates are corrected to use the translation tools.

DocBook is, for now, available in French and English - the browser
locale is loaded by default if it is one of those.

New page for 403 errors - excpetions now forward the request to the
correct route (a method of the "DocBookController").